# PERF: improve iloc list indexing