--- a/pandas/core/series.py
+++ b/pandas/core/series.py
@@ -2378,7 +2378,7 @@ class Series(base.IndexOpsMixin, strings.StringAccessorMixin,
         --------
         numpy.ndarray.take
         """
-        nv.validate_take(tuple(), kwargs)
+        #nv.validate_take(tuple(), kwargs)
 
         # check/convert indicies here
         if convert:
@@ -2388,7 +2388,7 @@ class Series(base.IndexOpsMixin, strings.StringAccessorMixin,
         new_index = self.index.take(indices)
         new_values = self._values.take(indices)
         return self._constructor(new_values,
-                                 index=new_index).__finalize__(self)
+                                 index=new_index, fastpath=True).__finalize__(self)
 
     def isin(self, values):
         """