@jorisvandenbossche 2017-03-29
@jreback Can you take a look at this? 
@gfyoung do you have an opinion about the kwarg validation in `take`? In this case, this is always unnecessary, and thus adds overhead that could be avoided (one option is to have a `_take` that does not do this for internal use, but this makes the code more ugly)