@gfyoung 2017-03-29
@jorisvandenbossche : It's not ideal, but given that users have had issues with this in the past, I would prefer that you keep that in the code than remove it.  If you leave the check, what is the speedup?