@gfyoung 2017-03-29
@jorisvandenbossche : Pre-checking `kwargs` or `args` is certainly a possibility.