@jreback 2017-03-30
do we have micro benchmarks for this?