@jorisvandenbossche 2017-03-30
> do we have micro benchmarks for this?

We actually already do have some, but I should run them (at the top are only some basic `%timeit` timings)