@jreback 2017-03-31
@jorisvandenbossche yeah the ``fastpath=True`` does not re-infer things. So that seems to be more correct. This just highlites that preserving numpy semantics is hard, while have easy-to-use pandas ones.

If you can separate out this specific case (to a separate) test (and put a reference to this issue) would be great.

After you merge this, I think I will create some ``pandas2`` markers in pytest and start adding here.