@jreback 2017-04-01
@jorisvandenbossche I think travis was flaky, can you push again